fix: return type

+++ Final_Codes/Src/scene_game.h
@@ -8,8 +8,8 @@
 
 Scene scene_main_create(void);
 /*
-int get_power_up_timer_tick();
-int get_power_up_duration();
+int64_t get_power_up_timer_tick();
+int64_t get_power_up_duration();
 */
 
 #endif
